requirements.txt
Django==5.2.13
djangorestframework==3.17.1
django-cors-headers==4.9.0
django-filter==25.2
psycopg2-binary==2.9.11
celery==5.6.3
redis==5.3.0
gunicorn==23.0.0
whitenoise==6.9.0
python-dotenv==1.0.1
dj-database-url==2.3.0
